refactor: Split payment-order grpc_service.go into logical modules
Split the 1772-line grpc_service.go into focused modules:
- grpc_service.go (384 lines): Core types, interfaces, constants, constructors
- grpc_initiate.go (289 lines): InitiatePaymentOrder, storeIdempotencyFailure
- grpc_update.go (356 lines): UpdatePaymentOrder, gateway callback handlers
- grpc_lifecycle.go (523 lines): Cancel, Reverse, fail, reverseLedgerPosting
- grpc_query.go (83 lines): RetrievePaymentOrder, ListPaymentOrders
- grpc_mappers.go (219 lines): Proto/domain converters, publishEvent, helpers

Internal reorganization only. No API changes. All tests pass.
