+++ services/payment-order/service/grpc_service.go
@@ -7,33 +7,20 @@ import (
 	"fmt"
 	"log/slog"
 	"os"
-	"strings"
 	"time"
 
-	"github.com/google/uuid"
-	commonpb "github.com/meridianhub/meridian/api/proto/meridian/common/v1"
 	currentaccountv1 "github.com/meridianhub/meridian/api/proto/meridian/current_account/v1"
-	eventsv1 "github.com/meridianhub/meridian/api/proto/meridian/events/v1"
 	financialaccountingv1 "github.com/meridianhub/meridian/api/proto/meridian/financial_accounting/v1"
 	internalbankaccountv1 "github.com/meridianhub/meridian/api/proto/meridian/internal_bank_account/v1"
 	pb "github.com/meridianhub/meridian/api/proto/meridian/payment_order/v1"
 	"github.com/meridianhub/meridian/services/payment-order/adapters/gateway"
 	"github.com/meridianhub/meridian/services/payment-order/adapters/persistence"
 	"github.com/meridianhub/meridian/services/payment-order/config"
-	"github.com/meridianhub/meridian/services/payment-order/domain"
-	poobservability "github.com/meridianhub/meridian/services/payment-order/observability"
 	sharedclients "github.com/meridianhub/meridian/shared/pkg/clients"
 	"github.com/meridianhub/meridian/shared/pkg/idempotency"
-	"github.com/meridianhub/meridian/shared/pkg/proto/mappers"
 	"github.com/meridianhub/meridian/shared/platform/defaults"
 	"github.com/meridianhub/meridian/shared/platform/observability"
-	"github.com/meridianhub/meridian/shared/platform/tenant"
-	"github.com/samber/lo"
-	"google.golang.org/genproto/googleapis/type/money"
-	"google.golang.org/grpc/codes"
-	"google.golang.org/grpc/status"
 	"google.golang.org/protobuf/proto"
-	"google.golang.org/protobuf/types/known/timestamppb"
 )
 
 // Service errors
@@ -394,1379 +381,4 @@ func NewServiceWithConfig(cfg Config) (*Service, error) {
 		orchestrator:              orchestrator,
 		internalClearingEnabled:   cfg.InternalClearingEnabled,
 	}, nil
-}
-
-// InitiatePaymentOrder creates a new payment order and begins the saga.
-// nolint:gocognit // Complexity justified by validation, idempotency handling (TOCTOU), event publishing, and saga startup
-func (s *Service) InitiatePaymentOrder(ctx context.Context, req *pb.InitiatePaymentOrderRequest) (*pb.InitiatePaymentOrderResponse, error) {
-	start := time.Now()
-	defer func() {
-		s.logger.Info("initiate_payment_order completed",
-			"duration_ms", time.Since(start).Milliseconds())
-	}()
-
-	// Extract or generate correlation ID
-	correlationID := req.CorrelationId
-	if correlationID == "" {
-		correlationID = uuid.New().String()
-		s.logger.Info("generated correlation ID", "correlation_id", correlationID)
-	}
-
-	// Get idempotency key (required)
-	var idempotencyKey string
-	if req.IdempotencyKey != nil {
-		idempotencyKey = req.IdempotencyKey.Key
-	}
-	if idempotencyKey == "" {
-		return nil, status.Error(codes.InvalidArgument, "idempotency_key is required")
-	}
-	if len(idempotencyKey) > s.maxIdempotencyKeyLength {
-		return nil, status.Errorf(codes.InvalidArgument, "idempotency_key exceeds maximum length of %d", s.maxIdempotencyKeyLength)
-	}
-
-	// Check Redis idempotency FIRST (before database check) - provides distributed lock protection
-	tenantID, _ := tenant.FromContext(ctx)
-	idempKey := idempotency.Key{
-		TenantID:  string(tenantID),
-		Namespace: idempotencyNamespace,
-		Operation: "initiate",
-		EntityID:  req.DebtorAccountId, // Use account as entity scope
-		RequestID: idempotencyKey,
-	}
-
-	result, err := s.idempotencyService.Check(ctx, idempKey)
-	// If operation already processed, return cached result
-	if errors.Is(err, idempotency.ErrOperationAlreadyProcessed) && result != nil && result.Data != nil {
-		var cachedResp pb.InitiatePaymentOrderResponse
-		unmarshalErr := proto.Unmarshal(result.Data, &cachedResp)
-		if unmarshalErr == nil {
-			s.logger.Info("returning cached initiate result from Redis",
-				"payment_order_id", cachedResp.PaymentOrder.PaymentOrderId,
-				"idempotency_key", idempotencyKey)
-			poobservability.RecordIdempotentRequest("initiate_payment_order_redis")
-			return &cachedResp, nil
-		}
-		s.logger.Warn("failed to unmarshal cached idempotency result, falling back to database check",
-			"error", unmarshalErr)
-	} else if err != nil && !errors.Is(err, idempotency.ErrResultNotFound) {
-		s.logger.Error("idempotency check failed", "error", err)
-		return nil, status.Error(codes.Internal, "failed to check idempotency")
-	}
-
-	// Mark operation as pending (distributed lock to prevent concurrent duplicates).
-	// Note: If this succeeds but the operation later fails with a transient error (e.g., DB unavailable),
-	// the pending lock remains until TTL expires (5 minutes). This is intentional - it prevents
-	// concurrent retry attempts from creating duplicates. The client should retry after TTL expiry.
-	if err := s.idempotencyService.MarkPending(ctx, idempKey, idempotencyPendingTTL); err != nil {
-		s.logger.Error("failed to mark operation pending", "error", err)
-		return nil, status.Error(codes.Internal, "failed to acquire idempotency lock")
-	}
-
-	// Check for existing payment order with same idempotency key (idempotent).
-	// Note: This check has a TOCTOU race window where concurrent requests with the same
-	// idempotency key could both pass this check. The database unique constraint on
-	// idempotency_key is the authoritative guard - concurrent inserts will fail with a
-	// constraint violation and should be handled by returning the existing record.
-	existingPO, err := s.repo.FindByIdempotencyKey(ctx, idempotencyKey)
-	if err != nil && !errors.Is(err, persistence.ErrPaymentOrderNotFound) {
-		s.logger.Error("failed to check idempotency", "error", err)
-		// Don't cache internal errors - allow retry on recovery
-		return nil, status.Error(codes.Internal, "failed to check idempotency")
-	}
-	if existingPO != nil {
-		s.logger.Info("returning existing payment order (idempotent)",
-			"payment_order_id", existingPO.ID.String(),
-			"idempotency_key", idempotencyKey)
-		return &pb.InitiatePaymentOrderResponse{
-			PaymentOrder: toProto(existingPO),
-		}, nil
-	}
-
-	// Validate and convert amount
-	amount, err := protoToMoney(req.Amount)
-	if err != nil {
-		// Store validation failure for idempotent error responses
-		s.storeIdempotencyFailure(ctx, idempKey, fmt.Sprintf("invalid amount: %v", err))
-		return nil, status.Errorf(codes.InvalidArgument, "invalid amount: %v", err)
-	}
-	if !amount.IsPositive() {
-		s.storeIdempotencyFailure(ctx, idempKey, "amount must be positive")
-		return nil, status.Error(codes.InvalidArgument, "amount must be positive")
-	}
-
-	// Create domain payment order
-	po, err := domain.NewPaymentOrder(
-		req.DebtorAccountId,
-		req.CreditorReference,
-		amount,
-		idempotencyKey,
-		correlationID,
-	)
-	if err != nil {
-		s.logger.Error("failed to create payment order", "error", err)
-		s.storeIdempotencyFailure(ctx, idempKey, fmt.Sprintf("failed to create payment order: %v", err))
-		return nil, status.Errorf(codes.InvalidArgument, "failed to create payment order: %v", err)
-	}
-
-	// Persist to database
-	if err := s.repo.Create(ctx, po); err != nil {
-		// Handle idempotency key conflict (TOCTOU race): another request won the race
-		// Reload and return the existing payment order for idempotent behavior
-		if errors.Is(err, persistence.ErrIdempotencyKeyConflict) {
-			existingPO, findErr := s.repo.FindByIdempotencyKey(ctx, idempotencyKey)
-			if findErr != nil {
-				s.logger.Error("failed to retrieve existing payment order after idempotency conflict",
-					"error", findErr,
-					"idempotency_key", idempotencyKey)
-				// Don't cache internal errors - allow retry on recovery
-				return nil, status.Error(codes.Internal, "failed to retrieve payment order")
-			}
-			s.logger.Info("returning existing payment order (idempotency race)",
-				"payment_order_id", existingPO.ID.String(),
-				"idempotency_key", idempotencyKey)
-			return &pb.InitiatePaymentOrderResponse{
-				PaymentOrder: toProto(existingPO),
-			}, nil
-		}
-		s.logger.Error("failed to save payment order", "error", err)
-		// Don't cache internal errors - allow retry on recovery
-		return nil, status.Error(codes.Internal, "failed to save payment order")
-	}
-
-	s.logger.Info("payment order created",
-		"payment_order_id", po.ID.String(),
-		"debtor_account_id", po.DebtorAccountID,
-		"amount_cents", safeMinorUnits(po.Amount),
-		"currency", domain.CurrencyCode(po.Amount),
-		"idempotency_key", po.IdempotencyKey,
-		"correlation_id", correlationID)
-
-	// Publish PaymentOrderInitiated event to Kafka
-	// Publish event (publishEvent handles nil kafkaPublisher)
-	s.publishEvent(ctx, TopicPaymentOrderInitiated, po.ID.String(), &eventsv1.PaymentOrderInitiatedEvent{
-		EventId:           uuid.New().String(),
-		PaymentOrderId:    po.ID.String(),
-		DebtorAccountId:   po.DebtorAccountID,
-		CreditorReference: po.CreditorReference,
-		Amount:            toMoneyAmount(po.Amount),
-		CorrelationId:     po.CorrelationID,
-		CausationId:       po.ID.String(), // Initial event caused by the order creation
-		Timestamp:         timestamppb.Now(),
-		Version:           int64(po.Version),
-		IdempotencyKey:    po.IdempotencyKey,
-	})
-
-	// Convert to proto BEFORE starting the async goroutine to avoid data race
-	// The saga may modify po while toProto reads from it
-	responseProto := toProto(po)
-
-	// Store successful result in Redis for future idempotency checks
-	response := &pb.InitiatePaymentOrderResponse{PaymentOrder: responseProto}
-	responseData, marshalErr := proto.Marshal(response)
-	if marshalErr == nil {
-		storeErr := s.idempotencyService.StoreResult(ctx, idempotency.Result{
-			Key:         idempKey,
-			Status:      idempotency.StatusCompleted,
-			Data:        responseData,
-			CompletedAt: time.Now(),
-			TTL:         idempotencyResultTTL,
-		})
-		if storeErr != nil {
-			s.logger.Error("failed to store idempotency result", "error", storeErr)
-			// Continue - operation succeeded, caching is optimization
-		}
-	} else {
-		s.logger.Error("failed to marshal response for idempotency cache", "error", marshalErr)
-	}
-
-	// Use tenantID from idempotency check above
-	hasTenant := tenantID != ""
-
-	// Start saga orchestration asynchronously
-	// The saga runs in the background after returning the response
-	// nolint:contextcheck // Intentionally using background context for async saga orchestration
-	go func(paymentOrderID uuid.UUID, tid tenant.TenantID, hasTenantCtx bool) {
-		// Recover from panics to prevent silent goroutine termination
-		defer func() {
-			if r := recover(); r != nil {
-				s.logger.Error("panic in payment saga orchestration",
-					"panic", r,
-					"payment_order_id", paymentOrderID.String(),
-					"correlation_id", correlationID)
-				// Reload fresh state before failing - the original po may be stale
-				// if the saga made state transitions before panicking
-				failCtx := context.Background()
-				if hasTenantCtx {
-					failCtx = tenant.WithTenant(failCtx, tid)
-				}
-				freshPO, err := s.repo.FindByID(failCtx, paymentOrderID)
-				if err != nil {
-					s.logger.Error("failed to reload payment order after panic",
-						"payment_order_id", paymentOrderID.String(),
-						"error", err)
-					return
-				}
-				// Async path: log and swallow error - best effort failure handling
-				if err := s.failPaymentOrder(failCtx, freshPO, "internal panic during saga orchestration", "INTERNAL_ERROR"); err != nil {
-					s.logger.Error("failed to mark payment order as failed after panic",
-						"payment_order_id", paymentOrderID.String(),
-						"error", err)
-				}
-			}
-		}()
-		// Create saga context with timeout to prevent indefinite hangs
-		sagaCtx := context.Background()
-		if hasTenantCtx {
-			sagaCtx = tenant.WithTenant(sagaCtx, tid)
-		}
-		sagaCtx, cancel := context.WithTimeout(sagaCtx, s.sagaTimeout)
-		defer cancel()
-		if s.tracer != nil {
-			sagaCtx = observability.WithCorrelationID(sagaCtx, correlationID)
-		}
-
-		// Reload fresh state to avoid race with caller who may still reference po
-		freshPO, err := s.repo.FindByID(sagaCtx, paymentOrderID)
-		if err != nil {
-			s.logger.Error("failed to reload payment order for saga",
-				"payment_order_id", paymentOrderID.String(),
-				"error", err)
-			return
-		}
-		s.orchestrator.Orchestrate(sagaCtx, freshPO)
-	}(po.ID, tenantID, hasTenant)
-
-	// Prevent accidental access to po after goroutine launch - the goroutine
-	// reloads fresh state from DB, so any access to po here would be stale
-	po = nil //nolint:ineffassign // Intentional: prevent future code from accidentally using stale po
-
-	return &pb.InitiatePaymentOrderResponse{
-		PaymentOrder: responseProto,
-	}, nil
-}
-
-// storeIdempotencyFailure stores a failure result for idempotency tracking.
-// This is called when a request fails validation or processing, allowing
-// subsequent retries with the same idempotency key to receive the same error.
-func (s *Service) storeIdempotencyFailure(ctx context.Context, key idempotency.Key, errorMsg string) {
-	result := idempotency.Result{
-		Key:         key,
-		Status:      idempotency.StatusFailed,
-		Error:       errorMsg,
-		CompletedAt: time.Now(),
-		TTL:         idempotencyResultTTL,
-	}
-	if err := s.idempotencyService.StoreResult(ctx, result); err != nil {
-		s.logger.Warn("failed to store idempotency failure result",
-			"error", err,
-			"idempotency_key", key.String())
-	}
-}
-
-// failPaymentOrder handles payment order failure with proper state transition and event publishing.
-// Returns an error if the state transition or persistence fails. Callers in synchronous paths
-// (e.g., UpdatePaymentOrder) should propagate this error to clients. Callers in async paths
-// (e.g., saga orchestration) may log and swallow the error.
-//
-// Compensation logic:
-// - Reverses ledger entries if LedgerBookingID exists (defensive - normally empty on failure)
-// - Releases lien if payment was in RESERVED or EXECUTING state
-func (s *Service) failPaymentOrder(ctx context.Context, po *domain.PaymentOrder, reason string, errorCode string) error {
-	// Capture original status before transitioning (for event)
-	failedAtStatus := po.Status
-
-	// Check if lien needs to be released before transitioning
-	needsLienRelease := po.RequiresLienRelease()
-	lienID := po.LienID
-	originalLedgerBookingID := po.LedgerBookingID
-
-	// Reverse ledger posting if it exists (defensive - normally empty before COMPLETED)
-	// This handles edge cases where ledger posting succeeded but state transition failed
-	if originalLedgerBookingID != "" {
-		_, err := s.reverseLedgerPosting(ctx, po, fmt.Sprintf("Payment failure: %s", reason))
-		if err != nil {
-			// Log but continue - ledger reversal failure shouldn't block payment failure
-			// The orphaned booking log will be flagged for reconciliation
-			s.logger.Error("failed to reverse ledger posting on payment failure",
-				"error", err,
-				"payment_order_id", po.ID.String(),
-				"original_ledger_booking_id", originalLedgerBookingID)
-		}
-	}
-
-	// Transition to FAILED
-	if err := po.Fail(reason, errorCode); err != nil {
-		s.logger.Error("failed to transition to FAILED state",
-			"error", err,
-			"payment_order_id", po.ID.String())
-		return fmt.Errorf("failed to transition to FAILED state: %w", err)
-	}
-
-	if err := s.repo.Update(ctx, po); err != nil {
-		s.logger.Error("failed to persist FAILED state",
-			"error", err,
-			"payment_order_id", po.ID.String())
-		return fmt.Errorf("failed to persist FAILED state: %w", err)
-	}
-
-	// Release lien if needed
-	if needsLienRelease && lienID != "" && s.currentAccountClient != nil {
-		_, err := s.currentAccountClient.TerminateLien(ctx, &currentaccountv1.TerminateLienRequest{
-			LienId: lienID,
-			Reason: fmt.Sprintf("Payment order %s failed: %s", po.ID.String(), reason),
-		})
-		if err != nil {
-			s.logger.Error("failed to release lien after failure",
-				"error", err,
-				"lien_id", lienID,
-				"payment_order_id", po.ID.String())
-		}
-	}
-
-	// Publish PaymentOrderFailed event
-	s.publishEvent(ctx, TopicPaymentOrderFailed, po.ID.String(), &eventsv1.PaymentOrderFailedEvent{
-		EventId:         uuid.New().String(),
-		PaymentOrderId:  po.ID.String(),
-		DebtorAccountId: po.DebtorAccountID,
-		Amount:          toMoneyAmount(po.Amount),
-		FailureReason:   reason,
-		ErrorCode:       errorCode,
-		FailedAtStatus:  mapStatusToProto(failedAtStatus),
-		LienId:          lienID,
-		CorrelationId:   po.CorrelationID,
-		CausationId:     po.ID.String(),
-		Timestamp:       timestamppb.Now(),
-		Version:         int64(po.Version),
-		IdempotencyKey:  po.IdempotencyKey,
-	})
-
-	s.logger.Info("payment order failed",
-		"payment_order_id", po.ID.String(),
-		"reason", reason,
-		"error_code", errorCode,
-		"original_ledger_booking_id", originalLedgerBookingID,
-		"idempotency_key", po.IdempotencyKey,
-		"correlation_id", po.CorrelationID)
-
-	return nil
-}
-
-// RetrievePaymentOrder gets payment order details by ID.
-func (s *Service) RetrievePaymentOrder(ctx context.Context, req *pb.RetrievePaymentOrderRequest) (*pb.RetrievePaymentOrderResponse, error) {
-	// Parse payment order ID
-	poID, err := uuid.Parse(req.PaymentOrderId)
-	if err != nil {
-		return nil, status.Errorf(codes.InvalidArgument, "invalid payment order ID: %v", err)
-	}
-
-	// Retrieve from repository
-	po, err := s.repo.FindByID(ctx, poID)
-	if err != nil {
-		if errors.Is(err, persistence.ErrPaymentOrderNotFound) {
-			return nil, status.Errorf(codes.NotFound, "payment order not found: %s", req.PaymentOrderId)
-		}
-		s.logger.Error("failed to retrieve payment order", "error", err)
-		return nil, status.Error(codes.Internal, "failed to retrieve payment order")
-	}
-
-	return &pb.RetrievePaymentOrderResponse{
-		PaymentOrder: toProto(po),
-	}, nil
-}
-
-// UpdatePaymentOrder handles asynchronous gateway callbacks.
-// Implements idempotency, audit logging, and observability per task 11 requirements.
-// nolint:gocognit // Gateway callback handling requires multiple state transitions and idempotency checks
-func (s *Service) UpdatePaymentOrder(ctx context.Context, req *pb.UpdatePaymentOrderRequest) (*pb.UpdatePaymentOrderResponse, error) {
-	start := time.Now()
-	operationStatus := opStatusSuccess
-	gatewayStatusStr := req.GatewayStatus.String()
-
-	defer func() {
-		elapsed := time.Since(start)
-		poobservability.RecordOperationDuration("update_payment_order", operationStatus, elapsed)
-		poobservability.RecordGatewayCallback(gatewayStatusStr, operationStatus)
-		s.logger.Info("update_payment_order completed",
-			"duration_ms", elapsed.Milliseconds(),
-			"gateway_status", gatewayStatusStr,
-			"result", operationStatus)
-	}()
-
-	// Get idempotency key from webhook request (required per proto)
-	var webhookIdempotencyKey string
-	if req.IdempotencyKey != nil {
-		webhookIdempotencyKey = req.IdempotencyKey.Key
-	}
-	if webhookIdempotencyKey == "" {
-		operationStatus = opStatusError
-		return nil, status.Error(codes.InvalidArgument, "idempotency_key is required for webhook callbacks")
-	}
-
-	// Lookup payment order by ID or gateway reference ID
-	po, err := s.lookupPaymentOrder(ctx, req)
-	if err != nil {
-		operationStatus = opStatusError
-		return nil, err
-	}
-
-	// Check Redis idempotency for webhook (prevents duplicate processing)
-	tenantID, _ := tenant.FromContext(ctx)
-	idempKey := idempotency.Key{
-		TenantID:  string(tenantID),
-		Namespace: idempotencyNamespace,
-		Operation: "update",
-		EntityID:  po.ID.String(), // Use payment order ID as entity
-		RequestID: webhookIdempotencyKey,
-	}
-
-	idempResult, err := s.idempotencyService.Check(ctx, idempKey)
-	// If operation already processed, return cached result
-	if errors.Is(err, idempotency.ErrOperationAlreadyProcessed) && idempResult != nil && idempResult.Data != nil {
-		var cachedResp pb.UpdatePaymentOrderResponse
-		unmarshalErr := proto.Unmarshal(idempResult.Data, &cachedResp)
-		if unmarshalErr == nil {
-			s.logger.Info("returning cached update result from Redis",
-				"payment_order_id", po.ID.String(),
-				"idempotency_key", webhookIdempotencyKey)
-			operationStatus = opStatusIdempotent
-			poobservability.RecordIdempotentRequest("update_payment_order_redis")
-			return &cachedResp, nil
-		}
-		s.logger.Warn("failed to unmarshal cached idempotency result, continuing with normal processing",
-			"error", unmarshalErr)
-	} else if err != nil && !errors.Is(err, idempotency.ErrResultNotFound) {
-		s.logger.Error("idempotency check failed", "error", err)
-		operationStatus = opStatusError
-		return nil, status.Error(codes.Internal, "failed to check idempotency")
-	}
-
-	// Mark operation as pending (distributed lock)
-	if err := s.idempotencyService.MarkPending(ctx, idempKey, idempotencyPendingTTL); err != nil {
-		s.logger.Error("failed to mark webhook operation pending", "error", err)
-		operationStatus = opStatusError
-		return nil, status.Error(codes.Internal, "failed to acquire webhook idempotency lock")
-	}
-
-	s.logger.Info("processing gateway callback",
-		"payment_order_id", po.ID.String(),
-		"gateway_reference_id", po.GatewayReferenceID,
-		"current_status", po.Status,
-		"gateway_status", gatewayStatusStr,
-		"correlation_id", po.CorrelationID)
-
-	// Process based on gateway status
-	var response *pb.UpdatePaymentOrderResponse
-
-	switch req.GatewayStatus {
-	case pb.GatewayStatus_GATEWAY_STATUS_SETTLED:
-		result, err := s.handleSettledStatus(ctx, po)
-		if err != nil {
-			operationStatus = opStatusError
-			// Don't cache handler errors - may contain transient failures
-			return nil, err
-		}
-		if result.isIdempotent {
-			operationStatus = opStatusIdempotent
-		}
-		response = &pb.UpdatePaymentOrderResponse{PaymentOrder: toProto(result.po)}
-
-	case pb.GatewayStatus_GATEWAY_STATUS_REJECTED:
-		result, err := s.handleRejectedStatus(ctx, po, req.GatewayMessage)
-		if err != nil {
-			operationStatus = opStatusError
-			// Don't cache handler errors - may contain transient failures
-			return nil, err
-		}
-		if result.isIdempotent {
-			operationStatus = opStatusIdempotent
-		}
-		response = &pb.UpdatePaymentOrderResponse{PaymentOrder: toProto(result.po)}
-
-	case pb.GatewayStatus_GATEWAY_STATUS_PENDING:
-		if err := s.handlePendingStatus(po, req.GatewayReferenceId); err != nil {
-			operationStatus = opStatusError
-			// Don't cache handler errors - may contain transient failures
-			return nil, err
-		}
-		response = &pb.UpdatePaymentOrderResponse{PaymentOrder: toProto(po)}
-
-	case pb.GatewayStatus_GATEWAY_STATUS_UNSPECIFIED:
-		operationStatus = opStatusError
-		s.storeIdempotencyFailure(ctx, idempKey, "gateway status is required")
-		return nil, status.Error(codes.InvalidArgument, "gateway status is required")
-
-	default:
-		operationStatus = opStatusError
-		s.storeIdempotencyFailure(ctx, idempKey, fmt.Sprintf("unknown gateway status: %v", req.GatewayStatus))
-		return nil, status.Errorf(codes.InvalidArgument, "unknown gateway status: %v", req.GatewayStatus)
-	}
-
-	// Store successful result in Redis for future idempotency checks
-	responseData, marshalErr := proto.Marshal(response)
-	if marshalErr == nil {
-		storeErr := s.idempotencyService.StoreResult(ctx, idempotency.Result{
-			Key:         idempKey,
-			Status:      idempotency.StatusCompleted,
-			Data:        responseData,
-			CompletedAt: time.Now(),
-			TTL:         idempotencyResultTTL,
-		})
-		if storeErr != nil {
-			s.logger.Error("failed to store webhook idempotency result", "error", storeErr)
-			// Continue - operation succeeded, caching is optimization
-		}
-	} else {
-		s.logger.Error("failed to marshal webhook response for idempotency cache", "error", marshalErr)
-	}
-
-	return response, nil
-}
-
-// lookupPaymentOrder finds a payment order by ID or gateway reference ID.
-func (s *Service) lookupPaymentOrder(ctx context.Context, req *pb.UpdatePaymentOrderRequest) (*domain.PaymentOrder, error) {
-	var po *domain.PaymentOrder
-	var err error
-
-	if req.PaymentOrderId != "" {
-		poID, parseErr := uuid.Parse(req.PaymentOrderId)
-		if parseErr != nil {
-			return nil, status.Errorf(codes.InvalidArgument, "invalid payment order ID: %v", parseErr)
-		}
-		po, err = s.repo.FindByID(ctx, poID)
-	} else if req.GatewayReferenceId != "" {
-		po, err = s.repo.FindByGatewayReferenceID(ctx, req.GatewayReferenceId)
-	} else {
-		return nil, status.Error(codes.InvalidArgument, "either payment_order_id or gateway_reference_id must be provided")
-	}
-
-	if err != nil {
-		if errors.Is(err, persistence.ErrPaymentOrderNotFound) {
-			return nil, status.Error(codes.NotFound, "payment order not found")
-		}
-		s.logger.Error("failed to find payment order", "error", err)
-		return nil, status.Error(codes.Internal, "failed to find payment order")
-	}
-
-	return po, nil
-}
-
-// updateResult holds the result of a status update operation.
-type updateResult struct {
-	po           *domain.PaymentOrder
-	isIdempotent bool
-}
-
-// handleSettledStatus processes a SETTLED gateway callback.
-// Implements idempotency: returns success if already COMPLETED.
-// Posts double-entry ledger journal entries BEFORE completing the payment.
-// nolint:gocognit // Payment completion requires ledger posting, state transition, and event publishing
-func (s *Service) handleSettledStatus(ctx context.Context, po *domain.PaymentOrder) (*updateResult, error) {
-	// Idempotency check: if already completed, return success without modification
-	if po.Status == domain.PaymentOrderStatusCompleted {
-		s.logger.Info("idempotent settled callback - payment already completed",
-			"payment_order_id", po.ID.String(),
-			"correlation_id", po.CorrelationID)
-		poobservability.RecordIdempotentRequest("update_payment_order_settled")
-		return &updateResult{po: po, isIdempotent: true}, nil
-	}
-
-	// Post ledger entries BEFORE completing the payment order
-	// This ensures double-entry bookkeeping is recorded before the payment is marked complete
-	ledgerBookingID, err := s.orchestrator.PostLedgerEntries(ctx, po)
-	if err != nil {
-		s.logger.Error("failed to post ledger entries",
-			"payment_order_id", po.ID.String(),
-			"error", err)
-		// Mark the payment as FAILED if ledger posting fails
-		if failErr := s.failPaymentOrder(ctx, po, fmt.Sprintf("ledger posting failed: %v", err), "LEDGER_POSTING_FAILED"); failErr != nil {
-			s.logger.Error("failed to mark payment as failed after ledger posting failure",
-				"payment_order_id", po.ID.String(),
-				"error", failErr)
-		}
-		return nil, status.Errorf(codes.Internal, "failed to post ledger entries: %v", err)
-	}
-
-	// Attempt state transition with the ledger booking ID
-	if err := po.Complete(ledgerBookingID); err != nil {
-		if errors.Is(err, domain.ErrInvalidPaymentOrderTransition) {
-			return nil, status.Errorf(codes.FailedPrecondition,
-				"cannot complete payment order in %s state: %v", po.Status, err)
-		}
-		return nil, status.Errorf(codes.Internal, "failed to complete payment: %v", err)
-	}
-
-	// Mark lien execution as pending before saving
-	if po.LienID != "" {
-		po.SetLienExecutionPending()
-	}
-
-	// Persist state change
-	if err := s.repo.Update(ctx, po); err != nil {
-		s.logger.Error("failed to update payment order to COMPLETED",
-			"error", err,
-			"payment_order_id", po.ID.String())
-		return nil, status.Error(codes.Internal, "failed to update payment order")
-	}
-
-	// Record metrics
-	poobservability.RecordCompletion(domain.CurrencyCode(po.Amount))
-	poobservability.RecordPaymentAmount(domain.CurrencyCode(po.Amount), "completed", safeMinorUnits(po.Amount))
-
-	// Execute lien asynchronously with retry mechanism
-	// The lien execution status is tracked in the payment order for reconciliation
-	if s.currentAccountClient != nil && po.LienID != "" {
-		// Start async retry goroutine - this won't block the webhook response
-		// We create a new background context since the request context will be cancelled
-		// after the webhook response is sent
-		asyncCtx := context.Background()
-		if tenantID, hasTenant := tenant.FromContext(ctx); hasTenant {
-			asyncCtx = tenant.WithTenant(asyncCtx, tenantID)
-		}
-		go s.orchestrator.ExecuteLienWithRetry(asyncCtx, po.ID, po.LienID) //nolint:contextcheck // Async operation outlives request context
-	}
-
-	// Publish PaymentOrderCompleted event
-	s.publishEvent(ctx, TopicPaymentOrderCompleted, po.ID.String(), &eventsv1.PaymentOrderCompletedEvent{
-		EventId:            uuid.New().String(),
-		PaymentOrderId:     po.ID.String(),
-		DebtorAccountId:    po.DebtorAccountID,
-		Amount:             toMoneyAmount(po.Amount),
-		LienId:             po.LienID,
-		GatewayReferenceId: po.GatewayReferenceID,
-		LedgerBookingId:    po.LedgerBookingID,
-		CorrelationId:      po.CorrelationID,
-		CausationId:        po.ID.String(),
-		Timestamp:          timestamppb.Now(),
-		Version:            int64(po.Version),
-		IdempotencyKey:     po.IdempotencyKey,
-	})
-
-	// Audit log for successful completion
-	s.logger.Info("payment order completed via gateway callback",
-		"payment_order_id", po.ID.String(),
-		"gateway_reference_id", po.GatewayReferenceID,
-		"ledger_booking_id", ledgerBookingID,
-		"amount_cents", safeMinorUnits(po.Amount),
-		"currency", domain.CurrencyCode(po.Amount),
-		"lien_id", po.LienID,
-		"idempotency_key", po.IdempotencyKey,
-		"correlation_id", po.CorrelationID)
-
-	return &updateResult{po: po, isIdempotent: false}, nil
-}
-
-// handleRejectedStatus processes a REJECTED gateway callback.
-// Implements idempotency: returns success if already FAILED.
-func (s *Service) handleRejectedStatus(ctx context.Context, po *domain.PaymentOrder, gatewayMessage string) (*updateResult, error) {
-	// Idempotency check: if already failed, return success without modification
-	if po.Status == domain.PaymentOrderStatusFailed {
-		s.logger.Info("idempotent rejected callback - payment already failed",
-			"payment_order_id", po.ID.String(),
-			"correlation_id", po.CorrelationID)
-		poobservability.RecordIdempotentRequest("update_payment_order_rejected")
-		return &updateResult{po: po, isIdempotent: true}, nil
-	}
-
-	// Fail the payment - synchronous path: propagate error to client
-	if err := s.failPaymentOrder(ctx, po, gatewayMessage, "GATEWAY_REJECTED"); err != nil {
-		return nil, status.Errorf(codes.Internal, "failed to mark payment as rejected: %v", err)
-	}
-
-	// Record metrics after successful persistence to ensure accuracy
-	poobservability.RecordRejection(domain.CurrencyCode(po.Amount), poobservability.ErrorCategoryGatewayRejected)
-	poobservability.RecordPaymentAmount(domain.CurrencyCode(po.Amount), "rejected", safeMinorUnits(po.Amount))
-
-	// Audit log for rejection
-	s.logger.Info("payment order rejected via gateway callback",
-		"payment_order_id", po.ID.String(),
-		"gateway_reference_id", po.GatewayReferenceID,
-		"gateway_message", gatewayMessage,
-		"amount_cents", safeMinorUnits(po.Amount),
-		"currency", domain.CurrencyCode(po.Amount),
-		"lien_id", po.LienID,
-		"idempotency_key", po.IdempotencyKey,
-		"correlation_id", po.CorrelationID)
-
-	return &updateResult{po: po, isIdempotent: false}, nil
-}
-
-// handlePendingStatus processes a PENDING gateway callback.
-// Validates state and logs - no state transition needed.
-func (s *Service) handlePendingStatus(po *domain.PaymentOrder, gatewayRefID string) error {
-	// Validate that we're still in EXECUTING state - PENDING callbacks for
-	// terminal states (COMPLETED, FAILED, etc.) should be rejected as stale
-	if po.Status != domain.PaymentOrderStatusExecuting {
-		return status.Errorf(codes.FailedPrecondition,
-			"cannot process PENDING callback: payment order is in %s state", po.Status)
-	}
-
-	// No state change needed - still waiting for final confirmation
-	s.logger.Info("payment still pending at gateway",
-		"payment_order_id", po.ID.String(),
-		"gateway_reference_id", gatewayRefID,
-		"correlation_id", po.CorrelationID)
-
-	return nil
-}
-
-// CancelPaymentOrder cancels a payment order before completion.
-func (s *Service) CancelPaymentOrder(ctx context.Context, req *pb.CancelPaymentOrderRequest) (*pb.CancelPaymentOrderResponse, error) {
-	// Validate cancellation reason - required for audit purposes
-	if req.CancellationReason == "" {
-		return nil, status.Error(codes.InvalidArgument, "cancellation_reason is required")
-	}
-
-	// Parse payment order ID
-	poID, err := uuid.Parse(req.PaymentOrderId)
-	if err != nil {
-		return nil, status.Errorf(codes.InvalidArgument, "invalid payment order ID: %v", err)
-	}
-
-	// Retrieve payment order
-	po, err := s.repo.FindByID(ctx, poID)
-	if err != nil {
-		if errors.Is(err, persistence.ErrPaymentOrderNotFound) {
-			return nil, status.Errorf(codes.NotFound, "payment order not found: %s", req.PaymentOrderId)
-		}
-		return nil, status.Error(codes.Internal, "failed to retrieve payment order")
-	}
-
-	// Check if already cancelled (idempotent)
-	if po.Status == domain.PaymentOrderStatusCancelled {
-		return &pb.CancelPaymentOrderResponse{PaymentOrder: toProto(po)}, nil
-	}
-
-	// Check if can be cancelled
-	if !po.CanCancel() {
-		return nil, status.Errorf(codes.FailedPrecondition,
-			"payment order cannot be cancelled in status %s", po.Status)
-	}
-
-	// Check if lien needs to be released
-	needsLienRelease := po.RequiresLienRelease()
-	lienID := po.LienID
-
-	// Cancel the payment order
-	if err := po.Cancel(req.CancellationReason); err != nil {
-		return nil, status.Errorf(codes.Internal, "failed to cancel payment order: %v", err)
-	}
-
-	if err := s.repo.Update(ctx, po); err != nil {
-		return nil, status.Error(codes.Internal, "failed to update payment order")
-	}
-
-	// Release lien if needed
-	if needsLienRelease && lienID != "" && s.currentAccountClient != nil {
-		_, termErr := s.currentAccountClient.TerminateLien(ctx, &currentaccountv1.TerminateLienRequest{
-			LienId: lienID,
-			Reason: fmt.Sprintf("Payment order %s cancelled: %s", po.ID.String(), req.CancellationReason),
-		})
-		if termErr != nil {
-			s.logger.Error("failed to release lien after cancellation",
-				"error", termErr,
-				"lien_id", lienID)
-			// Continue - cancellation succeeded, lien release can be retried
-		}
-	}
-
-	// Publish PaymentOrderCancelled event
-	s.publishEvent(ctx, TopicPaymentOrderCancelled, po.ID.String(), &eventsv1.PaymentOrderCancelledEvent{
-		EventId:            uuid.New().String(),
-		PaymentOrderId:     po.ID.String(),
-		DebtorAccountId:    po.DebtorAccountID,
-		Amount:             toMoneyAmount(po.Amount),
-		CancellationReason: req.CancellationReason,
-		CancelledBy:        req.CancelledBy,
-		LienId:             lienID,
-		CorrelationId:      po.CorrelationID,
-		CausationId:        po.ID.String(),
-		Timestamp:          timestamppb.Now(),
-		Version:            int64(po.Version),
-		IdempotencyKey:     po.IdempotencyKey,
-	})
-
-	s.logger.Info("payment order cancelled",
-		"payment_order_id", po.ID.String(),
-		"reason", req.CancellationReason,
-		"cancelled_by", req.CancelledBy,
-		"amount_cents", safeMinorUnits(po.Amount),
-		"currency", domain.CurrencyCode(po.Amount),
-		"idempotency_key", po.IdempotencyKey,
-		"correlation_id", po.CorrelationID)
-
-	return &pb.CancelPaymentOrderResponse{
-		PaymentOrder: toProto(po),
-	}, nil
-}
-
-// ListPaymentOrders returns a paginated list of payment orders.
-// Uses cursor-based pagination for consistent results even when items are inserted/deleted.
-// The cursor is an opaque token encoding (created_at, id) for deterministic ordering.
-func (s *Service) ListPaymentOrders(ctx context.Context, req *pb.ListPaymentOrdersRequest) (*pb.ListPaymentOrdersResponse, error) {
-	if req.DebtorAccountId == "" {
-		return nil, status.Error(codes.InvalidArgument, "debtor_account_id is required for listing")
-	}
-
-	// Parse and validate pagination parameters
-	pageSize := s.defaultPageSize
-	if req.Pagination != nil && req.Pagination.PageSize > 0 {
-		pageSize = int(req.Pagination.PageSize)
-		if pageSize > s.maxPageSize {
-			pageSize = s.maxPageSize
-		}
-	}
-
-	// Decode cursor from page token (empty token = first page)
-	var cursor persistence.Cursor
-	if req.Pagination != nil && req.Pagination.PageToken != "" {
-		var err error
-		cursor, err = persistence.DecodeCursor(req.Pagination.PageToken)
-		if err != nil {
-			return nil, status.Error(codes.InvalidArgument, "invalid page_token")
-		}
-	}
-
-	// Query with cursor-based pagination
-	result, err := s.repo.FindByDebtorAccountIDWithCursor(ctx, req.DebtorAccountId, pageSize, cursor)
-	if err != nil {
-		s.logger.Error("failed to list payment orders", "error", err)
-		return nil, status.Error(codes.Internal, "failed to list payment orders")
-	}
-
-	return &pb.ListPaymentOrdersResponse{
-		PaymentOrders: lo.Map(result.PaymentOrders, func(po *domain.PaymentOrder, _ int) *pb.PaymentOrder {
-			return toProto(po)
-		}),
-		Pagination: &commonpb.PaginationResponse{
-			NextPageToken: result.NextCursor,
-			TotalCount:    result.TotalCount,
-		},
-	}, nil
-}
-
-// ReversePaymentOrder reverses a completed payment order (post-completion compensation).
-// This creates compensating ledger entries and transitions the order to REVERSED.
-// Idempotent: returns success if already reversed.
-func (s *Service) ReversePaymentOrder(ctx context.Context, req *pb.ReversePaymentOrderRequest) (*pb.ReversePaymentOrderResponse, error) {
-	// Validate reversal reason - required for audit purposes
-	if req.ReversalReason == "" {
-		return nil, status.Error(codes.InvalidArgument, "reversal_reason is required")
-	}
-
-	// Validate reversed_by - required for audit purposes
-	if req.ReversedBy == "" {
-		return nil, status.Error(codes.InvalidArgument, "reversed_by is required")
-	}
-
-	// Parse payment order ID
-	poID, err := uuid.Parse(req.PaymentOrderId)
-	if err != nil {
-		return nil, status.Errorf(codes.InvalidArgument, "invalid payment order ID: %v", err)
-	}
-
-	// Check Redis idempotency for reversal
-	// Use payment_order_id as the request ID since each payment order can only be reversed once
-	tenantID, _ := tenant.FromContext(ctx)
-	idempKey := idempotency.Key{
-		TenantID:  string(tenantID),
-		Namespace: idempotencyNamespace,
-		Operation: "reverse",
-		EntityID:  req.PaymentOrderId,
-		RequestID: req.PaymentOrderId, // Payment order ID is the natural idempotency key for reversals
-	}
-
-	idempResult, err := s.idempotencyService.Check(ctx, idempKey)
-	// If operation already processed, return cached result
-	if errors.Is(err, idempotency.ErrOperationAlreadyProcessed) && idempResult != nil && idempResult.Data != nil {
-		var cachedResp pb.ReversePaymentOrderResponse
-		unmarshalErr := proto.Unmarshal(idempResult.Data, &cachedResp)
-		if unmarshalErr == nil {
-			s.logger.Info("returning cached reversal result from Redis",
-				"payment_order_id", req.PaymentOrderId)
-			poobservability.RecordIdempotentRequest("reverse_payment_order_redis")
-			return &cachedResp, nil
-		}
-		s.logger.Warn("failed to unmarshal cached idempotency result, falling back to database check",
-			"error", unmarshalErr)
-	} else if err != nil && !errors.Is(err, idempotency.ErrResultNotFound) {
-		s.logger.Error("idempotency check failed", "error", err)
-		return nil, status.Error(codes.Internal, "failed to check idempotency")
-	}
-
-	// Mark operation as pending (distributed lock)
-	if err := s.idempotencyService.MarkPending(ctx, idempKey, idempotencyPendingTTL); err != nil {
-		s.logger.Error("failed to mark reversal operation pending", "error", err)
-		return nil, status.Error(codes.Internal, "failed to acquire reversal idempotency lock")
-	}
-
-	// Retrieve payment order
-	po, err := s.repo.FindByID(ctx, poID)
-	if err != nil {
-		if errors.Is(err, persistence.ErrPaymentOrderNotFound) {
-			s.storeIdempotencyFailure(ctx, idempKey, fmt.Sprintf("payment order not found: %s", req.PaymentOrderId))
-			return nil, status.Errorf(codes.NotFound, "payment order not found: %s", req.PaymentOrderId)
-		}
-		// Don't cache internal errors - allow retry on recovery
-		return nil, status.Error(codes.Internal, "failed to retrieve payment order")
-	}
-
-	// Check if already reversed (idempotent)
-	if po.Status == domain.PaymentOrderStatusReversed {
-		return &pb.ReversePaymentOrderResponse{PaymentOrder: toProto(po)}, nil
-	}
-
-	// Check if can be reversed
-	if !po.CanReverse() {
-		s.storeIdempotencyFailure(ctx, idempKey, fmt.Sprintf("payment order cannot be reversed in status %s", po.Status))
-		return nil, status.Errorf(codes.FailedPrecondition,
-			"payment order cannot be reversed in status %s (only COMPLETED orders can be reversed)", po.Status)
-	}
-
-	// Store original ledger booking ID for the event
-	originalLedgerBookingID := po.LedgerBookingID
-
-	// Create compensating ledger entries if original posting exists
-	// This must happen before state transition to ensure ledger consistency
-	compensatingBookingID, err := s.reverseLedgerPosting(ctx, po, req.ReversalReason)
-	if err != nil {
-		s.logger.Error("failed to create compensating ledger entries for reversal",
-			"payment_order_id", po.ID.String(),
-			"original_ledger_booking_id", originalLedgerBookingID,
-			"error", err)
-		// Don't cache internal errors - allow retry on recovery
-		return nil, status.Errorf(codes.Internal, "failed to create compensating ledger entries: %v", err)
-	}
-
-	// Reverse the payment order
-	if err := po.Reverse(req.ReversalReason); err != nil {
-		// Don't cache internal errors - allow retry on recovery
-		return nil, status.Errorf(codes.Internal, "failed to reverse payment order: %v", err)
-	}
-
-	// Update in database
-	if err := s.repo.Update(ctx, po); err != nil {
-		// Don't cache internal errors - allow retry on recovery
-		return nil, status.Error(codes.Internal, "failed to update payment order")
-	}
-
-	// Publish PaymentOrderReversed event with compensating booking ID
-	s.publishEvent(ctx, TopicPaymentOrderReversed, po.ID.String(), &eventsv1.PaymentOrderReversedEvent{
-		EventId:                     uuid.New().String(),
-		PaymentOrderId:              po.ID.String(),
-		DebtorAccountId:             po.DebtorAccountID,
-		Amount:                      toMoneyAmount(po.Amount),
-		ReversalReason:              req.ReversalReason,
-		ReversedBy:                  req.ReversedBy,
-		OriginalLedgerBookingId:     originalLedgerBookingID,
-		CompensatingLedgerBookingId: compensatingBookingID,
-		CorrelationId:               po.CorrelationID,
-		CausationId:                 po.ID.String(),
-		Timestamp:                   timestamppb.Now(),
-		Version:                     int64(po.Version),
-		IdempotencyKey:              po.IdempotencyKey,
-	})
-
-	s.logger.Info("payment order reversed",
-		"payment_order_id", po.ID.String(),
-		"reason", req.ReversalReason,
-		"reversed_by", req.ReversedBy,
-		"amount_cents", safeMinorUnits(po.Amount),
-		"currency", domain.CurrencyCode(po.Amount),
-		"original_ledger_booking_id", originalLedgerBookingID,
-		"compensating_booking_id", compensatingBookingID,
-		"correlation_id", po.CorrelationID)
-
-	// Store successful result in Redis for future idempotency checks
-	response := &pb.ReversePaymentOrderResponse{PaymentOrder: toProto(po)}
-	responseData, marshalErr := proto.Marshal(response)
-	if marshalErr == nil {
-		storeErr := s.idempotencyService.StoreResult(ctx, idempotency.Result{
-			Key:         idempKey,
-			Status:      idempotency.StatusCompleted,
-			Data:        responseData,
-			CompletedAt: time.Now(),
-			TTL:         idempotencyResultTTL,
-		})
-		if storeErr != nil {
-			s.logger.Error("failed to store reversal idempotency result", "error", storeErr)
-			// Continue - operation succeeded, caching is optimization
-		}
-	} else {
-		s.logger.Error("failed to marshal reversal response for idempotency cache", "error", marshalErr)
-	}
-
-	return response, nil
-}
-
-// Helper functions
-
-// publishEvent publishes a Kafka event if the publisher is configured.
-// This is best-effort/fire-and-forget: errors are logged but not retried or persisted.
-// Failed events are NOT automatically recovered. For guaranteed delivery, consider
-// implementing a transactional outbox pattern in the future.
-func (s *Service) publishEvent(ctx context.Context, topic string, key string, event proto.Message) {
-	if s.kafkaPublisher == nil {
-		return
-	}
-	if err := s.kafkaPublisher.Publish(ctx, topic, key, event); err != nil {
-		s.logger.Error("failed to publish event",
-			"topic", topic,
-			"key", key,
-			"error", err)
-	} else {
-		s.logger.Info("published event",
-			"topic", topic,
-			"key", key)
-	}
-}
-
-// toProto converts a domain PaymentOrder to proto PaymentOrder
-func toProto(po *domain.PaymentOrder) *pb.PaymentOrder {
-	proto := &pb.PaymentOrder{
-		PaymentOrderId:        po.ID.String(),
-		DebtorAccountId:       po.DebtorAccountID,
-		CreditorReference:     po.CreditorReference,
-		Amount:                toMoneyAmount(po.Amount),
-		Status:                mapStatusToProto(po.Status),
-		LienId:                po.LienID,
-		GatewayReferenceId:    po.GatewayReferenceID,
-		CorrelationId:         po.CorrelationID,
-		IdempotencyKey:        po.IdempotencyKey,
-		FailureReason:         po.FailureReason,
-		CreatedAt:             timestamppb.New(po.CreatedAt),
-		UpdatedAt:             timestamppb.New(po.UpdatedAt),
-		Version:               int64(po.Version),
-		LedgerBookingId:       po.LedgerBookingID,
-		CausationId:           po.CausationID,
-		ErrorCode:             po.ErrorCode,
-		LienExecutionStatus:   mapLienExecutionStatusToProto(po.LienExecutionStatus),
-		LienExecutionAttempts: safeIntToInt32(po.LienExecutionAttempts),
-		LienExecutionError:    po.LienExecutionError,
-	}
-
-	// Set optional timestamps
-	if po.ReservedAt != nil {
-		proto.ReservedAt = timestamppb.New(*po.ReservedAt)
-	}
-	if po.ExecutingAt != nil {
-		proto.ExecutingAt = timestamppb.New(*po.ExecutingAt)
-	}
-	if po.CompletedAt != nil {
-		proto.CompletedAt = timestamppb.New(*po.CompletedAt)
-	}
-	if po.FailedAt != nil {
-		proto.FailedAt = timestamppb.New(*po.FailedAt)
-	}
-	if po.CancelledAt != nil {
-		proto.CancelledAt = timestamppb.New(*po.CancelledAt)
-	}
-	if po.ReversedAt != nil {
-		proto.ReversedAt = timestamppb.New(*po.ReversedAt)
-	}
-
-	return proto
-}
-
-// toMoneyAmount converts domain Money (in cents) to proto MoneyAmount (units + nanos).
-// The conversion splits cents into units (dollars) and nanos:
-//   - units = cents / 100 (integer division)
-//   - nanos = (cents % 100) * 10^7 (remainder converted to nanos)
-//
-// This is a lossless conversion since cents are exact representations.
-// The inverse operation protoToMoney uses symmetric rounding for any sub-cent precision.
-func toMoneyAmount(m domain.Money) *commonpb.MoneyAmount {
-	// Use domain.ToMinorUnits for the conversion
-	amountCents := domain.ToMinorUnits(m)
-	units := amountCents / 100
-	remainder := amountCents % 100
-	// #nosec G115 - remainder is always -99 to 99
-	nanos := int32(remainder * nanosPerCent)
-
-	return &commonpb.MoneyAmount{
-		Amount: &money.Money{
-			CurrencyCode: domain.CurrencyCode(m),
-			Units:        units,
-			Nanos:        nanos,
-		},
-	}
-}
-
-// safeMinorUnits converts Money to minor units (cents).
-// Used for logging and metrics where returning an error is not practical.
-// The new quantity package uses arbitrary precision, so overflow is not a concern.
-func safeMinorUnits(m domain.Money) int64 {
-	return domain.ToMinorUnits(m)
-}
-
-// protoToMoney converts proto MoneyAmount to domain Money
-func protoToMoney(amount *commonpb.MoneyAmount) (domain.Money, error) {
-	if amount == nil || amount.Amount == nil {
-		return domain.Money{}, ErrAmountRequired
-	}
-
-	// Validate nanos is within bounds (per Google Money spec)
-	const maxNanos = 999999999
-	if amount.Amount.Nanos < -maxNanos || amount.Amount.Nanos > maxNanos {
-		return domain.Money{}, ErrInvalidNanos
-	}
-
-	// Convert to cents with symmetric rounding (round half away from zero)
-	unitsCents := amount.Amount.Units * 100
-	var nanosCents int64
-	if amount.Amount.Nanos >= 0 {
-		nanosCents = int64((amount.Amount.Nanos + nanosRoundingOffset) / nanosPerCent)
-	} else {
-		nanosCents = int64((amount.Amount.Nanos - nanosRoundingOffset) / nanosPerCent)
-	}
-	totalCents := unitsCents + nanosCents
-
-	return domain.NewMoney(amount.Amount.CurrencyCode, totalCents)
-}
-
-// mapStatusToProto maps domain status to proto status
-func mapStatusToProto(status domain.PaymentOrderStatus) pb.PaymentOrderStatus {
-	switch status {
-	case domain.PaymentOrderStatusInitiated:
-		return pb.PaymentOrderStatus_PAYMENT_ORDER_STATUS_INITIATED
-	case domain.PaymentOrderStatusReserved:
-		return pb.PaymentOrderStatus_PAYMENT_ORDER_STATUS_RESERVED
-	case domain.PaymentOrderStatusExecuting:
-		return pb.PaymentOrderStatus_PAYMENT_ORDER_STATUS_EXECUTING
-	case domain.PaymentOrderStatusCompleted:
-		return pb.PaymentOrderStatus_PAYMENT_ORDER_STATUS_COMPLETED
-	case domain.PaymentOrderStatusFailed:
-		return pb.PaymentOrderStatus_PAYMENT_ORDER_STATUS_FAILED
-	case domain.PaymentOrderStatusCancelled:
-		return pb.PaymentOrderStatus_PAYMENT_ORDER_STATUS_CANCELLED
-	case domain.PaymentOrderStatusReversed:
-		return pb.PaymentOrderStatus_PAYMENT_ORDER_STATUS_REVERSED
-	default:
-		return pb.PaymentOrderStatus_PAYMENT_ORDER_STATUS_UNSPECIFIED
-	}
-}
-
-// mapLienExecutionStatusToProto maps domain LienExecutionStatus to proto
-func mapLienExecutionStatusToProto(status domain.LienExecutionStatus) pb.LienExecutionStatus {
-	switch status {
-	case domain.LienExecutionStatusUnspecified:
-		return pb.LienExecutionStatus_LIEN_EXECUTION_STATUS_UNSPECIFIED
-	case domain.LienExecutionStatusPending:
-		return pb.LienExecutionStatus_LIEN_EXECUTION_STATUS_PENDING
-	case domain.LienExecutionStatusSucceeded:
-		return pb.LienExecutionStatus_LIEN_EXECUTION_STATUS_SUCCEEDED
-	case domain.LienExecutionStatusFailed:
-		return pb.LienExecutionStatus_LIEN_EXECUTION_STATUS_FAILED
-	default:
-		return pb.LienExecutionStatus_LIEN_EXECUTION_STATUS_UNSPECIFIED
-	}
-}
-
-// safeIntToInt32 safely converts an int to int32 with bounds checking.
-// Returns math.MaxInt32 if the value exceeds int32 range.
-// This is used for lien execution attempts which should never exceed
-// a small number in practice, but we need safe conversion for gosec.
-func safeIntToInt32(n int) int32 {
-	const maxInt32 = 1<<31 - 1
-	if n > maxInt32 {
-		return maxInt32
-	}
-	if n < -maxInt32-1 {
-		return -maxInt32 - 1
-	}
-	return int32(n)
-}
-
-// reverseLedgerPosting creates reversal entries for a completed payment.
-// This is used for refunds and failed payment reversals.
-func (s *Service) reverseLedgerPosting(ctx context.Context, po *domain.PaymentOrder, reason string) (string, error) {
-	// No ledger entry to reverse if LedgerBookingID is empty
-	if po.LedgerBookingID == "" {
-		s.logger.Debug("no ledger entry to reverse - payment had no ledger posting",
-			"payment_order_id", po.ID.String())
-		return "", nil
-	}
-
-	// Skip ledger reversal if financial accounting client is not configured
-	// This allows the service to operate without FA integration for testing
-	if s.financialAccountingClient == nil || s.gatewayAccountConfig == nil {
-		s.logger.Warn("skipping ledger reversal - financial accounting client not configured",
-			"payment_order_id", po.ID.String(),
-			"ledger_booking_id", po.LedgerBookingID)
-		return "", nil
-	}
-
-	// Get the gateway contra-account from configuration
-	gatewayID := extractGatewayIDFromRef(po.GatewayReferenceID)
-	contraAccountID, err := s.gatewayAccountConfig.GetContraAccount(gatewayID)
-	if err != nil {
-		return "", fmt.Errorf("failed to get contra-account for gateway %s: %w", gatewayID, err)
-	}
-
-	// Convert domain currency to proto currency
-	currencyCode := domain.CurrencyCode(po.Amount)
-	protoCurrency := mappers.CurrencyCodeToProto(currencyCode)
-	if protoCurrency == commonpb.Currency_CURRENCY_UNSPECIFIED {
-		s.logger.Warn("unsupported currency for reversal posting",
-			"currency", currencyCode,
-			"payment_order_id", po.ID.String())
-		return "", fmt.Errorf("%w: %s", ErrUnsupportedCurrency, currencyCode)
-	}
-
-	// Step 1: Create a BookingLog in PENDING status for the reversal
-	reversalBookingLogIDempKey := fmt.Sprintf("reversal-booking-log-%s", po.IdempotencyKey)
-	bookingLogResp, err := s.financialAccountingClient.InitiateFinancialBookingLog(ctx, &financialaccountingv1.InitiateFinancialBookingLogRequest{
-		FinancialAccountType:    commonpb.AccountType_ACCOUNT_TYPE_CURRENT,
-		ProductServiceReference: "payment-order-reversal",
-		BusinessUnitReference:   "payment-order-service",
-		ChartOfAccountsRules:    "payment-reversal",
-		BaseCurrency:            protoCurrency,
-		IdempotencyKey: &commonpb.IdempotencyKey{
-			Key: reversalBookingLogIDempKey,
-		},
-	})
-	if err != nil {
-		return "", fmt.Errorf("failed to create reversal booking log: %w", err)
-	}
-	reversalBookingLogID := bookingLogResp.FinancialBookingLog.Id
-
-	s.logger.Debug("created reversal booking log",
-		"reversal_booking_log_id", reversalBookingLogID,
-		"original_booking_log_id", po.LedgerBookingID,
-		"payment_order_id", po.ID.String(),
-		"reason", reason)
-
-	// Convert amount from cents to google.type.Money format
-	amountCents := domain.ToMinorUnits(po.Amount)
-	postingAmount := &money.Money{
-		CurrencyCode: currencyCode,
-		Units:        amountCents / 100,
-		Nanos:        int32((amountCents % 100) * 10000000),
-	}
-	valueDate := timestamppb.Now()
-
-	// Step 2: Create CREDIT posting (customer account - funds returning)
-	// This reverses the original DEBIT on the customer account
-	reversalCreditIdempKey := fmt.Sprintf("reversal-credit-%s", po.IdempotencyKey)
-	_, err = s.financialAccountingClient.CaptureLedgerPosting(ctx, &financialaccountingv1.CaptureLedgerPostingRequest{
-		FinancialBookingLogId: reversalBookingLogID,
-		PostingDirection:      commonpb.PostingDirection_POSTING_DIRECTION_CREDIT,
-		PostingAmount:         postingAmount,
-		AccountId:             po.DebtorAccountID,
-		ValueDate:             valueDate,
-		IdempotencyKey: &commonpb.IdempotencyKey{
-			Key: reversalCreditIdempKey,
-		},
-	})
-	if err != nil {
-		s.logger.Error("RECONCILIATION_REQUIRED: reversal booking log orphaned after credit posting failure",
-			"reversal_booking_log_id", reversalBookingLogID,
-			"original_booking_log_id", po.LedgerBookingID,
-			"booking_log_status", "PENDING",
-			"payment_order_id", po.ID.String(),
-			"failed_step", "reversal_credit_posting",
-			"debtor_account", po.DebtorAccountID,
-			"error", err.Error())
-		return "", fmt.Errorf("failed to create reversal credit posting for account %s: %w", po.DebtorAccountID, err)
-	}
-
-	s.logger.Debug("created reversal credit posting",
-		"reversal_booking_log_id", reversalBookingLogID,
-		"account_id", po.DebtorAccountID,
-		"amount_cents", amountCents,
-		"payment_order_id", po.ID.String())
-
-	// Step 3: Create DEBIT posting (gateway contra-account - reducing liability)
-	// This reverses the original CREDIT on the gateway contra-account
-	reversalDebitIdempKey := fmt.Sprintf("reversal-debit-%s", po.IdempotencyKey)
-	_, err = s.financialAccountingClient.CaptureLedgerPosting(ctx, &financialaccountingv1.CaptureLedgerPostingRequest{
-		FinancialBookingLogId: reversalBookingLogID,
-		PostingDirection:      commonpb.PostingDirection_POSTING_DIRECTION_DEBIT,
-		PostingAmount:         postingAmount,
-		AccountId:             contraAccountID,
-		ValueDate:             valueDate,
-		IdempotencyKey: &commonpb.IdempotencyKey{
-			Key: reversalDebitIdempKey,
-		},
-	})
-	if err != nil {
-		s.logger.Error("RECONCILIATION_REQUIRED: reversal booking log orphaned after debit posting failure",
-			"reversal_booking_log_id", reversalBookingLogID,
-			"original_booking_log_id", po.LedgerBookingID,
-			"booking_log_status", "PENDING",
-			"payment_order_id", po.ID.String(),
-			"failed_step", "reversal_debit_posting",
-			"debtor_account", po.DebtorAccountID,
-			"contra_account", contraAccountID,
-			"has_credit_posting", true,
-			"error", err.Error())
-		return "", fmt.Errorf("failed to create reversal debit posting for account %s: %w", contraAccountID, err)
-	}
-
-	s.logger.Debug("created reversal debit posting",
-		"reversal_booking_log_id", reversalBookingLogID,
-		"account_id", contraAccountID,
-		"amount_cents", amountCents,
-		"payment_order_id", po.ID.String())
-
-	// Step 4: Update BookingLog status to POSTED
-	_, err = s.financialAccountingClient.UpdateFinancialBookingLog(ctx, &financialaccountingv1.UpdateFinancialBookingLogRequest{
-		Id:     reversalBookingLogID,
-		Status: commonpb.TransactionStatus_TRANSACTION_STATUS_POSTED,
-	})
-	if err != nil {
-		s.logger.Error("RECONCILIATION_REQUIRED: reversal booking log status update failed after successful postings",
-			"reversal_booking_log_id", reversalBookingLogID,
-			"original_booking_log_id", po.LedgerBookingID,
-			"booking_log_status", "PENDING",
-			"target_status", "POSTED",
-			"payment_order_id", po.ID.String(),
-			"failed_step", "reversal_status_update",
-			"has_credit_posting", true,
-			"has_debit_posting", true,
-			"resolution", "manually update booking log status to POSTED",
-			"error", err.Error())
-		return "", fmt.Errorf("failed to update reversal booking log to POSTED: %w", err)
-	}
-
-	s.logger.Info("reversal ledger posting completed successfully",
-		"reversal_booking_log_id", reversalBookingLogID,
-		"original_booking_log_id", po.LedgerBookingID,
-		"payment_order_id", po.ID.String(),
-		"debtor_account", po.DebtorAccountID,
-		"contra_account", contraAccountID,
-		"amount_cents", amountCents,
-		"currency", currencyCode,
-		"reason", reason)
-
-	return reversalBookingLogID, nil
-}
-
-// extractGatewayIDFromRef extracts the gateway identifier from a gateway reference ID.
-//
-// Gateway Reference ID Format:
-// Real gateways should use the format "{gateway_id}-{unique_reference}", e.g.:
-//   - "stripe-pm_1234abcd" -> returns "stripe"
-//   - "adyen-PSP-REF-123"  -> returns "adyen"
-//
-// Mock/Test gateways use special prefixes:
-//   - "GW-{uuid}"      -> returns "mock" (mock gateway format)
-//   - "gateway-{ref}"  -> returns "mock" (test helper format)
-//
-// Returns "unknown" for empty or invalid references.
-func extractGatewayIDFromRef(gatewayRefID string) string {
-	if gatewayRefID == "" {
-		return "unknown"
-	}
-
-	// Detect mock/test gateway patterns
-	if strings.HasPrefix(gatewayRefID, "GW-") || strings.HasPrefix(gatewayRefID, "gateway-") {
-		return "mock"
-	}
-
-	// For other gateways, extract the prefix before the first dash
-	parts := strings.SplitN(gatewayRefID, "-", 2)
-	if len(parts) > 0 && parts[0] != "" {
-		return strings.ToLower(parts[0])
-	}
-	return "unknown"
 }
